+++ app/src/main/java/com/macoscope/ketchuplunch/model/lunch/MealService.kt
@@ -6,7 +6,7 @@ import java.math.BigDecimal
 class MealService(val scriptClient: ScriptClient, val userName: String) {
 
 
-    fun getUserMeals(weekIndex: Int, dayIndex: Int): List<Meal> {
+    fun getUserMeals(weekIndex: Long, dayIndex: Int): List<Meal> {
 
         val parameters: List<Any> = listOf(userName.split("@").first(), weekIndex, dayIndex)
         val mealRawList = scriptClient.getDataFromApi<List<Map<String, Any>>>("meal", parameters)
